adjustments

+++ mmdet/ops/masked_conv/src/masked_conv2d_kernel.cu
@@ -18,8 +18,8 @@ __global__ void MaskedIm2colForward(const int n, const scalar_t *data_im,
                                     const int height, const int width,
                                     const int kernel_h, const int kernel_w,
                                     const int pad_h, const int pad_w,
-                                    const long *mask_h_idx,
-                                    const long *mask_w_idx, const int mask_cnt,
+                                    const int *mask_h_idx,
+                                    const int *mask_w_idx, const int mask_cnt,
                                     scalar_t *data_col) {
   // mask_cnt * channels
   CUDA_1D_KERNEL_LOOP(index, n) {
@@ -59,8 +59,8 @@ int MaskedIm2colForwardLaucher(const at::Tensor bottom_data, const int height,
   AT_DISPATCH_FLOATING_TYPES_AND_HALF(
       bottom_data.type(), "MaskedIm2colLaucherForward", ([&] {
         const scalar_t *bottom_data_ = bottom_data.data<scalar_t>();
-        const long *mask_h_idx_ = mask_h_idx.data<long>();
-        const long *mask_w_idx_ = mask_w_idx.data<long>();
+        const int *mask_h_idx_ = mask_h_idx.data<int>();
+        const int *mask_w_idx_ = mask_w_idx.data<int>();
         scalar_t *top_data_ = top_data.data<scalar_t>();
         MaskedIm2colForward<scalar_t>
             <<<GET_BLOCKS(output_size), THREADS_PER_BLOCK>>>(
@@ -74,8 +74,8 @@ int MaskedIm2colForwardLaucher(const at::Tensor bottom_data, const int height,
 template <typename scalar_t>
 __global__ void MaskedCol2imForward(const int n, const scalar_t *data_col,
                                     const int height, const int width,
-                                    const int channels, const long *mask_h_idx,
-                                    const long *mask_w_idx, const int mask_cnt,
+                                    const int channels, const int *mask_h_idx,
+                                    const int *mask_w_idx, const int mask_cnt,
                                     scalar_t *data_im) {
   CUDA_1D_KERNEL_LOOP(index, n) {
     const int m_index = index % mask_cnt;
@@ -99,8 +99,8 @@ int MaskedCol2imForwardLaucher(const at::Tensor bottom_data, const int height,
   AT_DISPATCH_FLOATING_TYPES_AND_HALF(
       bottom_data.type(), "MaskedCol2imLaucherForward", ([&] {
         const scalar_t *bottom_data_ = bottom_data.data<scalar_t>();
-        const long *mask_h_idx_ = mask_h_idx.data<long>();
-        const long *mask_w_idx_ = mask_w_idx.data<long>();
+        const int *mask_h_idx_ = mask_h_idx.data<int>();
+        const int *mask_w_idx_ = mask_w_idx.data<int>();
         scalar_t *top_data_ = top_data.data<scalar_t>();
 
         MaskedCol2imForward<scalar_t>
