+++ mmdet/ops/sigmoid_focal_loss/src/sigmoid_focal_loss_cuda.cu
@@ -23,7 +23,7 @@
 template <typename scalar_t>
 __global__ void SigmoidFocalLossForward(const int nthreads,
                                         const scalar_t *logits,
-                                        const long *targets,
+                                        const int *targets,
                                         const int num_classes,
                                         const float gamma, const float alpha,
                                         const int num, scalar_t *losses) {
@@ -60,7 +60,7 @@ __global__ void SigmoidFocalLossForward(const int nthreads,
 
 template <typename scalar_t>
 __global__ void SigmoidFocalLossBackward(
-    const int nthreads, const scalar_t *logits, const long *targets,
+    const int nthreads, const scalar_t *logits, const int *targets,
     const scalar_t *d_losses, const int num_classes, const float gamma,
     const float alpha, const int num, scalar_t *d_logits) {
   CUDA_1D_KERNEL_LOOP(i, nthreads) {
@@ -109,7 +109,7 @@ at::Tensor SigmoidFocalLoss_forward_cuda(const at::Tensor &logits,
   auto losses = at::empty({num_samples, logits.size(1)}, logits.options());
   auto losses_size = num_samples * logits.size(1);
 
-  dim3 grid(std::min(THCCeilDiv(losses_size, 512L), 4096L));
+  dim3 grid(std::min(THCCeilDiv((long)losses_size, 512L), 4096L));
   dim3 block(512);
 
   if (losses.numel() == 0) {
@@ -121,7 +121,7 @@ at::Tensor SigmoidFocalLoss_forward_cuda(const at::Tensor &logits,
       logits.type(), "SigmoidFocalLoss_forward", [&] {
         SigmoidFocalLossForward<scalar_t><<<grid, block>>>(
             losses_size, logits.contiguous().data<scalar_t>(),
-            targets.contiguous().data<long>(), num_classes, gamma, alpha,
+            targets.contiguous().data<int>(), num_classes, gamma, alpha,
             num_samples, losses.data<scalar_t>());
       });
   THCudaCheck(cudaGetLastError());
@@ -147,7 +147,7 @@ at::Tensor SigmoidFocalLoss_backward_cuda(const at::Tensor &logits,
   auto d_logits = at::zeros({num_samples, num_classes}, logits.options());
   auto d_logits_size = num_samples * logits.size(1);
 
-  dim3 grid(std::min(THCCeilDiv(d_logits_size, 512L), 4096L));
+  dim3 grid(std::min(THCCeilDiv((long)d_logits_size, 512L), 4096L));
   dim3 block(512);
 
   if (d_logits.numel() == 0) {
@@ -159,7 +159,7 @@ at::Tensor SigmoidFocalLoss_backward_cuda(const at::Tensor &logits,
       logits.type(), "SigmoidFocalLoss_backward", [&] {
         SigmoidFocalLossBackward<scalar_t><<<grid, block>>>(
             d_logits_size, logits.contiguous().data<scalar_t>(),
-            targets.contiguous().data<long>(),
+            targets.contiguous().data<int>(),
             d_losses.contiguous().data<scalar_t>(), num_classes, gamma, alpha,
             num_samples, d_logits.data<scalar_t>());
       });
